improved MTA:Lua memory usage and performance (introduced methods which only run if a method is of a certain trans); WIP: fix for C++ class integrity;

+++ Shared/sdk/CVector.h
@@ -104,7 +104,7 @@ public:
         vecRotation.fZ = atan2 ( fY, fX );
         CVector vecTemp ( sqrt ( fX * fX + fY * fY ), fZ, 0 );
         vecTemp.Normalize ();
-        vecRotation.fY = atan2 ( vecTemp.fX, vecTemp.fY ) - M_PI / 2;
+        vecRotation.fY = (float)atan2 ( vecTemp.fX, vecTemp.fY ) - (float)M_PI / 2;
         return vecRotation;
     }
